GaiA CONNECTED
- [ x ] Successfully connected to Gaia
- [ x ] Successfully write to Gaia
