+++ frontend/src/app/newUser/page.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
+import EditForm from '../../../components/EditForm'
 
 export default function newUser() {
   return (
     <div>
-      Hello
+      <EditForm></EditForm>
     </div>
   )
 }
